+++ qbraid/runtime/__init__.py
@@ -163,6 +163,11 @@ _lazy = {
         "OQCDevice",
         "OQCJob",
     ],
+    "rigetti": [
+        "RigettiProvider",
+        "RigettiDevice",
+        "RigettiJob",
+    ],
     "origin": [
         "OriginProvider",
         "OriginDevice",
@@ -220,6 +225,9 @@ if TYPE_CHECKING:
     from .quantinuum import QuantinuumDevice as QuantinuumDevice
     from .quantinuum import QuantinuumJob as QuantinuumJob
     from .quantinuum import QuantinuumProvider as QuantinuumProvider
+    from .rigetti import RigettiDevice as RigettiDevice
+    from .rigetti import RigettiJob as RigettiJob
+    from .rigetti import RigettiProvider as RigettiProvider
 
 
 def __getattr__(name):